+++ src/usage/tagged-union.ts
@@ -29,19 +29,31 @@ type AParam<Types extends AnyTypes> = {
   [k in keyof Types]: AType<Types[k]>
 }[keyof Types]
 
-export const makeTagged = <ProgURI extends ProgramsURI, InterpURI extends InterpretersURI>(
+type UnionTypes<
+  Types extends AnyTypes,
+  Tag extends keyof any,
+  ProgURI extends ProgramsURI,
+  InterpURI extends InterpretersURI
+> = {
+  [k in keyof Types]: M<EType<Types[k]>, AType<Types[k]> & { [t in Tag]: k }, ProgURI, InterpURI>
+}
+
+export function makeTagged<ProgURI extends ProgramsURI, InterpURI extends InterpretersURI>(
   summ: <A>(F: TaggedUnionProg<unknown, A, ProgURI>) => M<unknown, A, ProgURI, InterpURI>
-) => <Tag extends string>(tag: Tag) => <
-  Types extends {
-    [k in keyof Types]: M<EType<Types[k]>, AType<Types[k]> & { [t in Tag]: k }, ProgURI, InterpURI>
+) {
+  return <Tag extends string>(tag: Tag) => <
+    Types extends UnionTypes<Types, Tag, ProgURI, InterpURI>
+  >(
+    o: Types
+  ): ADTExt<unknown, AParam<Types>, TagType<Types>, ProgURI, InterpURI> => {
+    // Trust the outer signature - type lookup via unknown URI cannot have any semantic here
+    const summoned = summ<AParam<Types>>(
+      (F: any) =>
+        F.taggedUnion(
+          tag,
+          record.mapWithIndex((k, v: AnyM<ProgURI, InterpURI>) => (v as any)(F))(o)
+        ) // Trust
+    )
+    return summoned.tagged<TagType<Types>>(tag as typeof tag & TagType<Types>)(o as any) // We're bending reality as bit
   }
->(
-  o: Types
-): ADTExt<unknown, AParam<Types>, TagType<Types>, ProgURI, InterpURI> => {
-  // Trust the outer signature - type lookup via unknown URI cannot have any semantic here
-  const summoned = summ<AParam<Types>>(
-    (F: any) =>
-      F.taggedUnion(tag, record.mapWithIndex((k, v: AnyM<ProgURI, InterpURI>) => (v as any)(F))(o)) // Trust
-  )
-  return summoned.tagged<TagType<Types>>(tag as typeof tag & TagType<Types>)(o as any) // We're bending reality as bit
 }